test/requirements.txt
streamlit
litellm
python-dotenv
pypdf
faiss-cpu
